+++ CPU/InstructionDecode/Instructions/Flow/RtsInstruction.cs
@@ -5,7 +5,7 @@
     /// </summary>
     public class RtsInstruction : InstructionBase
     {
-        public RtsInstruction(ushort opCode, AddressingMode addressingMode, Mos6502Core core) : base("RTS", opCode, addressingMode, core)
+        public RtsInstruction(byte opCode, AddressingMode addressingMode, Mos6502Core core) : base("RTS", opCode, addressingMode, core)
         {
 
         }
